feat: unify process edit lcia results panel
